feat: now you can't rename or change the desc of your workspace
