+++ src/views/group/GroupViewNotesTab.tsx
@@ -9,67 +9,9 @@ import { useGetCurrentGroup } from '../../hooks/api/groups'
 import { useCreateNoteInCurrentGroup, useDeleteNoteInCurrentGroup, useListNotesInCurrentGroup } from '../../hooks/api/notes'
 import { FormatMessage, useOurIntl } from '../../i18n/TextComponent'
 import { V1Note } from '../../protorepo/openapi/typescript-axios'
+import NotesListGridItemContextMenu from '../notes/NotesListGridItemContextMenu'
 
-type NotesLiNotesListGridItemContextMenuProps = {
-  note: V1Note,
-  targetId: string,
-  options: {
-    name: string,
-    icon: (props: React.SVGProps<SVGSVGElement> & { title?: string | undefined; titleId?: string | undefined; }) => JSX.Element,
-    onClick: () => void,
-  }[],
-}
-
-const NotesListGridItemContextMenu: React.FC<NotesLiNotesListGridItemContextMenuProps> = (props) => {
-  const [contextData, setContextData] = React.useState({ visible:false, posX: 0, posY: 0})
-  const contextRef = React.useRef<HTMLDivElement>(null)
-
-  React.useEffect(() => {
-    const contextMenuEventHandler= (e) => {
-      const targetElement = document.getElementById(props.targetId)
-
-      if (targetElement && targetElement.contains(e.target)) {
-        e.preventDefault()
-        setContextData({ visible: true, posX: e.clientX, posY: e.clientY })
-      } else if(contextRef.current && !contextRef.current.contains(e.target)){
-        setContextData({ ...contextData, visible: false })
-      }
-    }
-
-    const offClickHandler= () => {
-      if(contextRef.current){
-        setContextData({ ...contextData, visible: false })
-      }
-    }
-
-    document.addEventListener('contextmenu', contextMenuEventHandler)
-    document.addEventListener('click', offClickHandler)
-    return () => {
-      document.removeEventListener('contextmenu', contextMenuEventHandler)
-      document.removeEventListener('click', offClickHandler)
-    }
-  }, [contextData, props.targetId])
-
-  return <div ref={contextRef} className='absolute w-48 list-none rounded-md border border-gray-200 bg-white shadow-lg' style={{ display:`${contextData.visible ? 'block' : 'none'}`, left: contextData.posX, top: contextData.posY }}>
-    <div className=''>
-      {props.options.map((option) => (
-        <li
-          key={`notes-list-grid-item-context-menu-${props.note.id}-${option.name}`}
-          className='group m-1 flex cursor-pointer items-center rounded-md px-3 py-2 text-sm hover:bg-gray-50'
-          onClick={() => {option.onClick()}}>
-          <div className='mr-2'>
-            {<option.icon className='h-5 w-5 stroke-2 text-gray-400 group-hover:text-gray-500' />}
-          </div>
-          <p className='text-gray-600 group-hover:text-black'>
-            {option.name}
-          </p>
-        </li>
-      ))}
-    </div>
-  </div>
-}
-
-export const NotesListGridItem: React.FC<{ note: V1Note }> = (props) => {
+const NotesListGridItem: React.FC<{ note: V1Note }> = (props) => {
   const authorQ = useGetAccount({accountId: props.note.authorAccountId})
   const deleteNoteQ = useDeleteNoteInCurrentGroup()
   const navigate = useNavigate()
@@ -120,7 +62,6 @@ const GroupViewNotesTab: React.FC = () => {
   return (
     <div className='grid w-full grid-rows-1 gap-4'>
       {/* Menu */}
-
       {/* Search bar */}
       <Stack direction='row' spacing={2}>
         <input
@@ -137,25 +78,31 @@ const GroupViewNotesTab: React.FC = () => {
           endIcon={
             createNoteQ.isLoading ?
               <ArrowPathIcon className='h-5 w-5 animate-spin text-blue-500' /> : 
-              <PlusIcon className='h-5 w-5 stroke-2 text-blue-500' />
+              <PlusIcon className='h-5 w-5 stroke-2' style={{color: '#2a777d' }} />
           }
         >
           <FormatMessage id='NOTE.newNote' />
         </Button>
       </Stack>
       {/* Notes Grid */}
-      <div className='grid grid-cols-2 gap-2 md:grid-cols-3 md:gap-3 lg:grid-cols-3 lg:gap-4 xl:grid-cols-4'>
-        {listNotesQ.isSuccess ? (
-          listNotesQ.data?.notes?.map((note, idx) => (
-            <NotesListGridItem
-              key={`group-view-notes-tab-grid-${note.id}-${idx}`}
-              note={note}
-            />
-          ))
-        ) : (
-          <div className='skeleton h-48 w-full' />
-        )}
-      </div>
+      {listNotesQ.isSuccess ? 
+        (listNotesQ.data.notes?.length !== 0 ? 
+          <div className='grid grid-cols-2 gap-2 md:grid-cols-3 md:gap-3 lg:grid-cols-3 lg:gap-4 xl:grid-cols-4'>
+            {listNotesQ.data.notes?.map((note, idx) => (
+              <NotesListGridItem
+                key={`group-view-notes-tab-grid-${note.id}-${idx}`}
+                note={note}
+              />
+            ))}
+          </div>
+          : 
+          <div className='flex h-full items-center justify-center lg:px-lg lg:pb-lg xl:px-xl xl:pb-xl'>
+            <div className='my-4 space-y-2 text-center'>
+              <FormatMessage id='NOTES.noNotes' />
+            </div>
+          </div>) :
+        <div className='skeleton h-48 w-full' />     
+      }
     </div>
   )
 }
