+++ src/views/home/HomeView.tsx
@@ -21,70 +21,72 @@ const HomeView: React.FC = () => {
   const listNotesQ = useListNotes({authorAccountId: authContext.accountId, limit: 8})
   const listGroupsQ = useListGroups({accountId: authContext.accountId, limit: 8})
 
-  return (<ViewSkeleton title={formatMessage({ id: 'GENERIC.home' })} panels={['group-activity-no-group']}>
-    <Container sx={{ mt: 3 }} >
-      <Box className='mb-4 rounded-3xl border border-gray-300'>
-        <Typography variant='h5' gutterBottom
-          sx={{ color: grey[800] }}
-          style={{
-            marginTop: '11.9px',
-            marginLeft: '20px'
-          }}>
-          <FormatMessage id='HOME.lastGroups' />
-        </Typography>
-        <div className='m-4 grid grid-cols-2 gap-2 md:grid-cols-3 md:gap-3 lg:grid-cols-3 lg:gap-4 xl:grid-cols-4'>
-          {listGroupsQ.isSuccess ? (
-            listGroupsQ.data?.groups?.map((group, idx) => {
-              const handleViewGroup = () => {
-                groupContext.changeGroup(group.id)
-                navigate(`/group/${group.id}`)
-              }
-              return (<div
-                className='flex h-24 w-full cursor-pointer flex-col items-center justify-center rounded-md border border-gray-100 bg-gray-50 p-2 transition-all hover:bg-gray-100 hover:shadow-inner'
-                onClick={handleViewGroup}
-                key={`group-view-notes-tab-grid-${group.id}-${idx}`}
-              >
-                <Group style={{ color: theme.palette.primary.main }} />
-                <div className='w-full text-center'>
-                  <div className='text-xs font-medium text-gray-800'>{group.name}</div>
-                </div>
-              </div>)
-            })
-          ) : (
-            <div className='skeleton h-48 w-full' />
-          )}
-        </div>
-      </Box>
-      <Box className='rounded-3xl border border-gray-300'>
-        <Typography variant='h5' gutterBottom
-          sx={{ color: grey[800] }}
-          style={{
-            marginTop: '11.9px',
-            marginLeft: '20px'
-          }}>
-          <FormatMessage id='HOME.lastNotes' />
-        </Typography>
-        {listNotesQ.isSuccess ? 
-          (listNotesQ.data.notes?.length !== 0 ? 
-            <div className='m-4 grid grid-cols-2 gap-2 md:grid-cols-3 md:gap-3 lg:grid-cols-3 lg:gap-4 xl:grid-cols-4'>
-              {listNotesQ.data?.notes?.map((note, idx) => (
-                <NotesListGridItemNoGroup
-                  key={`group-view-notes-tab-grid-${note.id}-${idx}`}
-                  note={note}
-                />
-              ))}
-            </div>
-            : 
-            <div className='flex h-full items-center justify-center lg:px-lg lg:pb-lg xl:px-xl xl:pb-xl'>
-              <div className='my-4 space-y-2 text-center'>
-                <FormatMessage id='NOTES.noNotes' />
+  return (
+    <ViewSkeleton title={formatMessage({ id: 'GENERIC.home' })} panels={['group-activity-no-group']}>
+      <Container sx={{ mt: 3 }} >
+        <Box className='mb-4 rounded-3xl border border-gray-300'>
+          <Typography variant='h5' gutterBottom
+            sx={{ color: grey[800] }}
+            style={{
+              marginTop: '11.9px',
+              marginLeft: '20px'
+            }}>
+            <FormatMessage id='HOME.lastGroups' />
+          </Typography>
+          <div className='m-4 grid grid-cols-2 gap-2 md:grid-cols-3 md:gap-3 lg:grid-cols-3 lg:gap-4 xl:grid-cols-4'>
+            {listGroupsQ.isSuccess ? (
+              listGroupsQ.data?.groups?.map((group, idx) => {
+                const handleViewGroup = () => {
+                  groupContext.changeGroup(group.id)
+                  navigate(`/group/${group.id}`)
+                }
+                return (<div
+                  className='flex h-24 w-full cursor-pointer flex-col items-center justify-center rounded-md border border-gray-100 bg-gray-50 p-2 transition-all hover:bg-gray-100 hover:shadow-inner'
+                  onClick={handleViewGroup}
+                  key={`group-view-notes-tab-grid-${group.id}-${idx}`}
+                >
+                  <Group style={{ color: theme.palette.primary.main }} />
+                  <div className='w-full text-center'>
+                    <div className='text-xs font-medium text-gray-800'>{group.name}</div>
+                  </div>
+                </div>)
+              })
+            ) : (
+              <div className='skeleton h-48 w-full' />
+            )}
+          </div>
+        </Box>
+        <Box className='rounded-3xl border border-gray-300'>
+          <Typography variant='h5' gutterBottom
+            sx={{ color: grey[800] }}
+            style={{
+              marginTop: '11.9px',
+              marginLeft: '20px'
+            }}>
+            <FormatMessage id='HOME.lastNotes' />
+          </Typography>
+          {listNotesQ.isSuccess ? 
+            (listNotesQ.data.notes?.length !== 0 ? 
+              <div className='m-4 grid grid-cols-2 gap-2 md:grid-cols-3 md:gap-3 lg:grid-cols-3 lg:gap-4 xl:grid-cols-4'>
+                {listNotesQ.data?.notes?.map((note, idx) => (
+                  <NotesListGridItemNoGroup
+                    key={`group-view-notes-tab-grid-${note.id}-${idx}`}
+                    note={note}
+                  />
+                ))}
               </div>
-            </div>) :
-          <div className='skeleton h-48 w-full' />     
-        }
-      </Box>
-    </Container>
-  </ViewSkeleton>)
+              : 
+              <div className='flex h-full items-center justify-center lg:px-lg lg:pb-lg xl:px-xl xl:pb-xl'>
+                <div className='my-4 space-y-2 text-center'>
+                  <FormatMessage id='NOTES.noNotes' />
+                </div>
+              </div>) :
+            <div className='skeleton h-48 w-full' />     
+          }
+        </Box>
+      </Container>
+    </ViewSkeleton>
+  )
 }
 
 export default HomeView
